Remove legacy version of the plugin
